server/requirements.txt
reportlab
Flask-CORS
Flask
python-dotenv
openai==0.28.0
PyPDF2
